Remove unneeded username lookups when saving new leaderboard entries
The names can still be logged, but only when Verbose logging is enabled. Otherwise, it will only log the number of entries that were added.
This dramatically improves the speed of adding new levels' leaderboard entries to the database
