+++ DistanceScraper/Utils.cs
@@ -1,5 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
+using System.Threading.Tasks;
+using SteamKit2;
 
 namespace DistanceScraper
 {
@@ -25,6 +29,34 @@ namespace DistanceScraper
 			Console.SetCursorPosition(0, Console.CursorTop);
 			Console.Write(new string(' ', Console.WindowWidth));
 			Console.SetCursorPosition(0, currentLineCursor);
+		}
+
+		public static async Task LogNewLeaderboardEntry(Leaderboard leaderboard, List<SteamUserStats.LeaderboardEntriesCallback.LeaderboardEntry> newEntries, Handlers handlers, BaseScraper scraper)
+		{
+			// Look for steamids that need caching
+			var steamIDs = new List<SteamID>();
+			foreach (var newEntry in newEntries)
+			{
+				var name = handlers.Friends.GetFriendPersonaName(newEntry.SteamID);
+				if (string.IsNullOrEmpty(name))
+				{
+					steamIDs.Add(newEntry.SteamID);
+				}
+			}
+
+			// Cache in batches
+			var steamIDBatches = steamIDs.Chunk(100);
+			foreach (var steamIDBatch in steamIDBatches)
+			{
+				await scraper.RequestUserInfo(steamIDBatch.ToList());
+			}
+
+			// Log information about the new time and who set it
+			foreach (var newEntry in newEntries)
+			{
+				var name = handlers.Friends.GetFriendPersonaName(newEntry.SteamID);
+				WriteLine($"New time: {name} set their first time on {leaderboard.LevelName}: {newEntry.Score / 1000.0:0.000}s with a rank of {newEntry.GlobalRank}!");
+			}
 		}
 	}
 }
